Diffusion Setup

Defining the observational scheme

To run the MCMC sampler we need to pass a setup object that defines the statistical model. In this section we will show how to define a diffusion model using a pre-defined DiffusionSetup. Nonetheless, be aware that the logic of the MCMC sampler is independent from the model specification, so it is possible to use the mcmc routines from this package routine to run it on other statistical models. For more information about these types of extensions see Generic MCMC.

Defining the processes

To define the DiffusionSetup we need to decide on the Target diffusion, an Auxiliary diffusion and the observation scheme. For instance, suppose that there are three observations:

obs = [...]
obs_times = [1.0, 2.0, 3.0]

Then we define the target diffusion globally and the auxiliary diffusion separately on each interval

P_target = TargetDiffusion(parameters)
P_auxiliary = [AuxiliaryDiffusion(parameters, o, t) for (o,t) in zip(obs, obs_times)]

To define the setup for partially observed diffusion it is enough to write:

model_setup = DiffusionSetup(P_target, P_auxiliary, PartObs()) # for first passage times use FPT()

BridgeSDEInference.DiffusionSetup — Type.

MCMCSetup

Groups together all objects that need to be passed to mcmc function from mcmc.jl.

source

Observations

To set the observations, apart from passing the observations and observation times it is necessary to pass the observational operators and as well as covariance of the noise. Additionally, one can pass additional information about the first passage time scheme.

L = ...
Σ = ...
set_observations!(model_setup, [L for _ in obs], [Σ for _ in obs], obs, obs_time)

BridgeSDEInference.set_observations! — Function.

set_observations!(setup::DiffusionSetup, Ls, Σs, obs, obs_times,
                  fpt=fill(nothing, length(obs)-1))

Store observations to setup. The observations follow the scheme V=LX+η, where V are the observations in obs, observed at times given in obs_times, L are the observation operators given in Ls, X is the unobserved, underlying diffusion and η is a Gaussian noise with mean 0 and covariance Σ with the last stored in Σs. fpt provides additional information in case the nature of observations has to do with first-passage times

source

Imputation grid

There are two objects that define the imputation grid. The time step dt and the time transformation that transforms a regular time-grid with equidistantly distributed imputation points. The second defaults to a usual transformation employed in papers on the guided proposals. [TO DO add also space-time transformation from the original paper for the bridges]. It is enough to call

dt = ...
set_imputation_grid!(model_setup, dt)

BridgeSDEInference.set_imputation_grid! — Function.

set_imputation_grid!(setup::DiffusionSetup, dt,
                     time_transf=(t₀,T) -> ((x) ->  t₀ + (x-t₀) * (2-(x-t₀)/(T-t₀))))

Define the imputation grid in setup. dt defines the granulatrity of the imputation grid and time_transf defines a time transformation to use for transforming equidistant grid.

source

Prior over the starting point

There are two types of priors for the starting point, either a delta hit at some specified value, corresponding to a known starting point and a Gaussian prior

BridgeSDEInference.KnownStartingPt — Type.

KnownStartingPt{T} <: StartingPtPrior

Indicates that the starting point is known and stores its value in y

source

BridgeSDEInference.GsnStartingPt — Type.

GsnStartingPt{T,S} <: StartingPtPrior

Indicates that the starting point is equipped with a Gaussian prior with mean μ and covariance matrix Σ. It also stores the most recently sampled white noise z used to compute the starting point and a precision matrix Λ:=Σ⁻¹. μ₀ and Σ₀ are the mean and covariance of the white noise

GsnStartingPt(μ::T, Σ::S)

Base constructor. It initialises the mean μ and covariance Σ parameters and Λ is set according to Λ:=Σ⁻¹

source

For instance, to set a known starting point it is enough to call:

set_x0_prior!(model_setup, KnownStartingPt(x0))

BridgeSDEInference.set_x0_prior! — Function.

set_priors!(setup::DiffusionSetup, priors, x0_prior, x0_guess=nothing)

Store the priors over parameters (in priors) and over the starting point (in x0_prior) into the object setup.

source

Auxiliary parameters

There are two auxiliary parameters that can be set by the user. The first one is a thinning factor for saving the paths that are sampled by the MCMC sampler. The second specifies an adaptation scheme for tuning Guided proposals [TODO change the latter].

Missing docstring.

Missing docstring for set_auxiliary!. Check Documenter's build log for details.

Initialisation of internal containers

Once all the setting functions above have been run (with the only exception set_auxiliary! being optional), i.e.

setup = DiffusionSetup(...)
set_observations!(setup, ...)
set_imputation_grid(setup, ...)
set_x0_prior!(setup, ...)

then the following function should be run

BridgeSDEInference.initialise! — Function.

initialise(::Type{K}, setup::DiffusionSetup; verbose=false)

Initialise the internal containers of setup. Check if all the necessary data has been passed to setup

source

Once run, the model setup is complete.