Setup

Defining the observational scheme and parameters of the Markov chain

There is a series of objects that define the Markov chain Monte Carlo sampler and that the user needs to define in order to be able to run the inference algorithm. To keep this processes structured an object MCMCSetup is defined. It allows for a systematic and concise way of defining the MCMC sampler.

Defining the processes

To define the MCMCSetup one needs to decide on the Target diffusion an Auxiliary diffusion and the observation scheme. For instance, suppose that there are three observations:

obs = [...]
obs_times = [1.0, 2.0, 3.0]

Then the target diffusion is defined globally and the auxiliary diffusion is defined separately on each interval

P_target = TargetDiffusion(parameters)
P_auxiliary = [AuxiliaryDiffusion(parameters, o, t) for (o,t) in zip(obs, obs_times)]

To define the setup for partially observed diffusion it is enough to write:

setup = MCMCSetup(P_target, P_auxiliary, PartObs()) # for first passage times use FPT()

BridgeSDEInference.MCMCSetup — Type.

MCMCSetup

Groups together all objects that need to be passed to mcmc function from mcmc.jl.

source

Observations

To set the observations, apart from passing the observations and observation times it is necessary to pass the observational operators and as well as covariance of the noise. Additionally, one can pass additional information about the first-passage scheme [TO DO add details on fpt].

L = ...
Σ = ...
set_observations!(setup, [L for _ in obs], [Σ for _ in obs], obs, obs_time)

BridgeSDEInference.set_observations! — Function.

set_observations!(setup::MCMCSetup, Ls, Σs, obs, obs_times,
                  fpt=fill(nothing, length(obs)-1))

Store observations to setup. The observations follow the scheme V=LX+η, where V are the observations in obs, observed at times given in obs_times, L are the observation operators given in Ls, X is the unobserved, underlying diffusion and η is a Gaussian noise with mean 0 and covariance Σ with the last stored in Σs. fpt provides additional information in case the nature of observations has to do with first-passage times

source

Imputation grid

There are two objects that define the imputation grid. The time step dt and the time transformation that transforms a regular time-grid with equidistantly distributed imputation points. The second defaults to a usual transformation employed in papers on the guided proposals. [TO DO add also space-time transformation from the original paper for the bridges]. It is enough to call

dt = ...
set_imputation_grid!(setup, dt)

BridgeSDEInference.set_imputation_grid! — Function.

set_imputation_grid!(setup::MCMCSetup, dt,
                     time_transf=(t₀,T) -> ((x) ->  t₀ + (x-t₀) * (2-(x-t₀)/(T-t₀))))

Define the imputation grid in setup. dt defines the granulatrity of the imputation grid and time_transf defines a time transformation to use for transforming equidistant grid.

source

Transition kernels

To define the updates of the parameters and the Wiener path a couple of objects need to specified. A boolean flag needs to be passed indicating whether any parameter updates are to be performed. If set to false then only the path is updated and the result it a marginal sampler on a path space. Additionally a memory (or persistence) parameter of the preconditioned Crank-Nicolson scheme needs to be set for the path updates. For the parameter updates three additional objects must be specified. A sequence of transition kernels–-one for each Gibbs step, a sequence of lists indicating parameters to be updated–-one list for each Gibbs step and a sequence of indicators about the types of parameter updates–-one for each Gibbs step. Additionally, an object describing an adaptation scheme for the auxiliary law can be passed [TODO add description to the last]

Random walk

The package provides an implementation of a random walk, which can be used as a generic transition kernel

BridgeSDEInference.RandomWalk — Type.

RandomWalk(ϵ::T, pos::S)

Defines a random walk on |ϵ|-dimensional space. `ϵ defines the maximal one-sided range of a single step and pos is a vector of indicators for whether a respective index is restricted to take only positive values. For elements restricted to take positive values, the update is done via: x⁽ⁿᵉʷ⁾ <- x⁽ᵒˡᵈ⁾eᵁ, where U ∼ Unif(-ϵ,ϵ). For unrestricted: x⁽ⁿᵉʷ⁾ <- x⁽ᵒˡᵈ⁾ + U, where U ∼ Unif(-ϵ,ϵ).

source

Indicators for parameter update

The indicators for parameter updates should be in a format of tuple of tuples (or arrays of arrays etc.). Each inner tuple corresponds to a single Gibbs step and the elements of the inner tuples give indices of parameters that are to be updated on a given Gibbs step. For instance: ((1,2,3),(5,)) says that in the first Gibbs step the first three parameters are to be updated, whereas in the second Gibbs step parameter 5 is to be updated.

Flags for the types of parameter updates

There are currently two different ways of updating parameters:

BridgeSDEInference.ConjugateUpdt — Type.

ConjugateUpdt <: ParamUpdateType

Type acting as a flag for update from full conditional (conjugate to a prior)

source

BridgeSDEInference.MetropolisHastingsUpdt — Type.

MetropolisHastingsUpdt <: ParamUpdateType

Flag for performing update according to Metropolis Hastings step

source

Setting transition kernels

An example of setting the transition kernels is as follows:

pCN = ... # memory paramter of the preconditioned Crank Nicolson scheme
update_parameters = true
set_transition_kernels!(setup,
                        [RandomWalk([],[]),
                         RandomWalk([3.0, 5.0, 5.0, 0.01, 0.5], 5)],
                        pCN, update_parameters, ((1,2,3),(5,)),
                        (ConjugateUpdt(),
                         MetropolisHastingsUpdt(),
                        ))

BridgeSDEInference.set_transition_kernels! — Function.

set_transition_kernels!(setup::MCMCSetup, transition_kernel,
                        crank_nicolson_memory=0.0, param_updt::Bool=true,
                        updt_coord=(Val((true,)),),
                        updt_type=(MetropolisHastingsUpdt(),),
                        adaptive_proposals=NoAdaptation())

Store the transition kernels for parameter update steps as well path imputation step in the setup object. ...

Arguments

setup: object to be set up

transition_kernel: collection of transition kernels (one for each param updt)

crank_nicolson_memory: memory parameter for random walk on a path space

param_updt: flag for whether to update parameters at all

updt_coord: collection indicating which coordinates to update

updt_type: collection indicating types of parameter updates

adaptive_proposals: object for adapting guided proposals

...

source

Priors

Priors need to be defined for each Gibbs step (or more precisely for each subset of parameters that is being updated). Additionally a prior over the starting position needs to be defined as well. A few convenience functions for the priors are provided by the package.

Prior over the starting point

There are two types of priors for the starting point, either a delta hit at some specified value, corresponding to a known starting point and a Gaussian prior

BridgeSDEInference.KnownStartingPt — Type.

KnownStartingPt{T} <: StartingPtPrior

Indicates that the starting point is known and stores its value in y

source

BridgeSDEInference.GsnStartingPt — Type.

GsnStartingPt{T,S} <: StartingPtPrior

Indicates that the starting point is equipped with a Gaussian prior with mean μ and covariance matrix Σ. It also stores the most recently sampled white noise z used to compute the starting point and a precision matrix Λ:=Σ⁻¹. μ₀ and Σ₀ are the mean and covariance of the white noise

GsnStartingPt(μ::T, Σ::S)

Base constructor. It initialises the mean μ and covariance Σ parameters and Λ is set according to Λ:=Σ⁻¹

source

Prior over the parameters

An improper flat prior is provided for quick tests:

BridgeSDEInference.ImproperPrior — Type.

ImproperPrior

Flat prior

source

Additionally, as each Gibbs step needs to have its own set of priors corresponding to the parameters being updated by this Gibbs step, the priors, similarly to indices of updated coordinates need to be grouped into tuples of tuples. To make this grouping easier on the user a Priors structure is provided.

BridgeSDEInference.Priors — Type.

Priors

Struct

struct Priors
    priors
    indicesForUpdt::Array{Array{Int64,1},1}
end

holds relevant information about priors: distributions that need to be used and the pattern in which MCMC makes use of those distributions. priors is a list of distributions with all priors that are to be used. indicesForUpdt is a list of lists–-the reason is as follows. Markov chains can have multiple transition kernels, each updating (possibly) multiple number of parameters (think of Gibbs sampler, which updates parameters in blocks)–-for each such update, priors for all updated parameters are needed. The inner list of indicesForUpdt gives indices of all priors that together (in a factories form) make up a prior for a respective transition kernel of the Markov chain

Priors(priors, indicesForUpdt)

Base constructor. priors is a list of priors, indicesForUpdt is a pattern of indices indicating in which way parameters are being updated

Priors(priors)

Convenience constructor. Most of the time each kernel of the Markov chain will be updating only a single parameter and thus only a single prior will be needed for each transition. In that case providing a list of priors in priors is sufficient. This constructor takes care of the internal objects in such setting.

source

Example 1

This is perhaps the most common use case. Suppose that n coordinates–-where we take n=3 for illustration purposes–-of prameter θ are being updated by the Markov chain. Suppose further that there are n transition kernels, each updating a separate coordinate. Let's assume for simplicity that each coordinate is equipped with an independent, improper prior. Then the Priors struct can be set up as follows:

n = 3
priors = Priors([ImproperPrior() for i in 1:n])

display(priors.priors)
(ImproperPrior(), ImproperPrior(), ImproperPrior())

display(priors.indicesForUpdt)
3-element Array{Array{Int64,1},1}:
 [1]
 [2]
 [3]

Notice that only a list of priors had to be supplied and the struct took care of setting up the indicesForUpdt container.

Example 2

Suppose that the Markov chain uses two transition kernels. The first transition kernel updates 3 coordinates of θ, which have a corresponding joint, multivariate Normal prior with some pre-specified covariance matrix Σ and mean 0. The second transition kernel updates 2 coordinates of θ, and these two are equipped with independent, improper priors. Then the Priors struct can be set up as follows:

using LinearAlgebra, Distributions
Σ = diagm(0=>[1000.0, 1000.0, 1000.0])
μ = [0.0,0.0,0.0]
priors = Priors([MvNormal(μ, Σ), ImproperPrior(), ImproperPrior()],
                [[1],[2,3]])
display(priors.priors)
(FullNormal(
dim: 3
μ: [0.0, 0.0, 0.0]
Σ: [1000.0 0.0 0.0; 0.0 1000.0 0.0; 0.0 0.0 1000.0]
)\mutable struct
  fields
end
, ImproperPrior(), ImproperPrior())

display(priors.indicesForUpdt)
2-element Array{Array{Int64,1},1}:
 [1]   
 [2, 3]

Notice that we had to pass all priors in a list and then specify that the first transition kernel uses only the first prior, whereas the second transition kernel uses two priors from the list.

Setting the priors

An example of setting the priors for two-step Gibbs sampler could be

μ, Σ, λ, Ω = ..., ..., ..., ...
set_priors!(setup, Priors((MvNormal(μ, Σ), ImproperPrior())), GsnStartingPt(λ, Ω), x0)

BridgeSDEInference.set_priors! — Function.

set_priors!(setup::MCMCSetup, priors, x0_prior, x0_guess=nothing)

Store the priors over parameters (in priors) and over the starting point (in x0_prior) into the object setup.

source

Blocking

Currently two choices of blocking are available:

Either no blocking at all, which is the default behaviour of set_blocking!

BridgeSDEInference.NoBlocking — Type.

struct NoBlocking <: BlockingSchedule end

Regular updates with no blocking

source

Or blocking using the chequerboard updating scheme.

ChequeredBlocking

For chequerboard updating scheme, at each observation a knot can be (but does not have to be) placed. IMPORTANT: The knot indexing starts at the first non-starting point observation. Suppose we have, say, 20 observations (excluding the starting point). Let's put a knot on every other observation, ending up with knots on observations with indices: [2,4,6,8,10,12,14,16,18,20]. Chequerboard updating scheme splits these knots into two, disjoint, interlaced subsets, i.e. [2,6,10,14,18] and [4,8,12,16,20]. This also splits the path into two interlaced sets of blocks: [1–2,3–6,7–10,11–14,15–18,19–20], [1–4,5–8,9–12,13–16,17–20] (where interval indexing starts with interval 1, whose end-points are the starting point and the first non-starting point observation). The path is updated in blocks. First, blocks [1–2,3–6,7–10,11–14,15–18,19–20] are updated conditionally on full and exact observations indexed with [2,6,10,14,18], as well as all the remaining, partial observations (indexed by [1,2,3,...,20]). Then, the other set of blocks is updated in the same manner. This is then repeated. To define the blocking behaviour, only the following needs to be written:

blocking = ChequeredBlocking()
blocking_params = (collect(2:20)[1:2:end], 10^(-10), SimpleChangePt(100))

The first defines the blocking updating scheme (in the future there might be a larger choice). The second line places the knots on [2,4,6,8,10,12,14,16,18,20]. Splitting into appropriate subsets is done internally. 10^(-10) is an artificial noise parameter that needs to be added for the numerical purposes. Ideally we want this to be as small as possible, however the algorithm may have problems with dealing with very small values. The last arguments aims to remedy this. SimpleChangePt(100) has two functions. One, it is a flag to the mcmc sampler that two sets of ODE solvers need to be employed: for the segment directly adjacent to a knot from the left ODE solvers for M⁺, L, μ are employed and H, Hν and c are computed as a by-product. On the remaining part of blocks, the ODE solvers for H, Hν and c are used directly. The second function of SimpleChangePt() is to indicate the point at which a change needs to be made between these two solvers (which for the example above is set to 100). The reason for this functionality is that solvers for M⁺, L, μ are more tolerant to very small values of the artificial noise.

To define an MCMC sampler with no blocking nothing needs to be done (it's a default). Alternatively, one can call

set_blocking!()

It resets the blocking to none. To pass the blocking scheme defined above one could call

set_blocking!(setup, blocking, blocking_params)

ODE solvers

There are a few standard, Runge-Kutta type ODE solvers implemented for solving the backward ODE systems defining M⁺, L, μ and H, Hν and c. These are:

Missing docstring.

Missing docstring for Ralston3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RK4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Tsit5. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Vern7. Check Documenter's build log for details.

Additionally, the change point between solvers for M⁺, L, μ and H, Hν and c can be set outside of blocking when setting the solvers. An example code is as follows

set_solver!(setup, Vern7(), NoChangePt())

BridgeSDEInference.set_solver! — Function.

set_solver!(setup::MCMCSetup, solver=Ralston3(), change_pt=NoChangePt())

Define the ODE solvers in setup. solver defines the type of ODE solver and change_pt gives information about the change point for switching between the solvers for M,L,μ and H,Hν,c

source

MCMC parameters

There are some additional parameters that need to be passed to an MCMC samplers. These need to be passed to

BridgeSDEInference.set_mcmc_params! — Function.

set_mcmc_params!(setup::MCMCSetup, num_mcmc_steps, save_iter=NaN,
                 verb_iter=NaN, skip_for_save=1, warm_up=0)

Define the parametrisation of the mcmc sampler. ...

Arguments

setup: object to be set up

num_mcmc_steps: total number of steps of the mcmc sampler

save_iter: save the path every ... iteration

verb_iter: print progress message to console every ... iteration

skip_for_save: when saving path, thin the grid by a factor of ...

uwarm_up: number of steps of the chain in which no param update is made

...

source

Initialisation of internal containers

Once all the setting functions above have been run (with the only exception set_blocking! being optional), i.e.

setup = MCMCSetup(...)
set_observations!(setup, ...)
set_imputation_grid(setup, ...)
set_transition_kernels!(setup, ...)
set_priors!(setup, ...)
set_mcmc_params!(setup, ...)
set_solver!(setup, ...)
set_blocking(setup, ...) # optionally

then the following function should be run

BridgeSDEInference.initialise! — Function.

initialise(::Type{K}, setup::MCMCSetup; verbose=false)

Initialise the internal containers of setup. Check if all the necessary data has been passed to setup

source

Once run, the setup is complete and it is possible to commence the MCMC sampler.