Querying the results

Querying the inference results

The output of running the mcmc function is a Workspace object.

BridgeSDEInference.Workspace — Type.

Workspace{ObsScheme,S,TX,TW,R,ST}

The main container of the mcmc function from mcmc.jl in which most data pertinent to sampling is stored

source

Querying the output can be done simply by calling the members of an instance of Workspace that is returned by the mcmc sampler. The task of writing suitable functions for this is left mainly to the user, we provide some generic plotting functions that can be used for testing here (not part of the package).